backend/requirements.txt
fastapi==0.99.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.1.0
alembic==1.11.1
psycopg-binary==3.2.0
pydantic==2.4.0
python-dotenv==1.1.1
passlib[bcrypt]==1.7.5
python-multipart==0.0.6
redis==4.9.0
aiofiles==23.2.1
